# Add color wheel to output