--- a/dogm/demo/main.cpp
+++ b/dogm/demo/main.cpp
@@ -254,14 +254,13 @@ std::vector<float2> load_measurement_from_image(const std::string &file_name)
     return meas_grid;
 }
 
-void createColorWheelImageAndMask(cv::Mat &destination_image, cv::Mat &destination_mask, const size_t hue_offset)
+cv::Mat createCircularColorGradient(const size_t hue_offset)
 {
     // Set linear gradient (180 levels). Needed because the OpenCV hue range is [0, 179], see
     // https://docs.opencv.org/3.2.0/df/d9d/tutorial_py_colorspaces.html
     const size_t hue_steps = 180;
     cv::Mat hue_image(hue_steps, hue_steps, CV_8UC3, cv::Scalar(0));
-    cv::Mat mask(hue_image.size(), CV_8UC3, cv::Scalar(0, 0, 0));
-
+    const cv::Point image_center{hue_image.cols / 2, hue_image.rows / 2};
     cv::Mat hsv{1, 1, CV_8UC3, cv::Scalar(0, 255, 255)};
     cv::Mat rgb{1, 1, CV_8UC3};
     for (int hue = 0; hue < hue_image.rows; hue++)
@@ -270,46 +269,63 @@ void createColorWheelImageAndMask(cv::Mat &destination_image, cv::Mat &destinati
         cv::cvtColor(hsv, rgb, cv::COLOR_HSV2RGB);
         hue_image.row(hue).setTo(rgb.at<cv::Vec3b>(0, 0));
     }
-
-    cv::linearPolar(hue_image, hue_image, cv::Point(hue_image.cols / 2, hue_image.rows / 2), 255,
+    cv::linearPolar(hue_image, hue_image, image_center, 255,
                     cv::INTER_CUBIC | cv::WARP_FILL_OUTLIERS | cv::WARP_INVERSE_MAP);
-
-    cv::circle(mask, cv::Point(mask.cols / 2, mask.rows / 2), hue_image.rows / 2, cv::Scalar(1, 1, 1), -1);
-    cv::circle(mask, cv::Point(mask.cols / 2, mask.rows / 2), hue_image.rows / 4, cv::Scalar(0, 0, 0), -1);
-    hue_image.copyTo(destination_image, mask);
-    mask.copyTo(destination_mask);
+    return hue_image;
 }
 
-void resizeImagesRelativeToShorterEdge(const cv::Mat &original_img, const float relative_size,
-                                       std::vector<cv::Mat> imgs)
+cv::Mat createCircleMask(const cv::Size2d &size)
 {
-    const float edge_length{std::min(original_img.cols, original_img.rows) * relative_size};
-    for (auto &img : imgs)
-    {
-        cv::resize(img, img, cv::Size(edge_length, edge_length));
-    }
+    cv::Mat mask(size, CV_8UC3, cv::Scalar(0, 0, 0));
+    const cv::Point image_center{mask.cols / 2, mask.rows / 2};
+    const float outer_radius = mask.rows / 2;
+    const float inner_radius = mask.rows / 4;
+    cv::circle(mask, image_center, outer_radius, cv::Scalar(1, 1, 1), -1);
+    cv::circle(mask, image_center, inner_radius, cv::Scalar(0, 0, 0), -1);
+    return mask;
 }
 
-void blendIntoImage(cv::Mat &img, cv::Mat &img_to_blend_in, const cv::Mat &alpha_mask_of_img_to_blend_in)
+cv::Mat createColorWheel(const size_t hue_offset = 0)
 {
-    cv::Rect roi = cv::Rect(img.cols - img_to_blend_in.cols, img.rows - img_to_blend_in.rows, img_to_blend_in.cols,
-                            img_to_blend_in.rows);
+    cv::Mat circular_color_gradient = createCircularColorGradient(hue_offset);
+    cv::Mat circle_mask = createCircleMask(circular_color_gradient.size());
+
+    cv::Mat color_wheel{};
+    circular_color_gradient.copyTo(color_wheel, circle_mask);
+    return color_wheel;
+}
+
+void resizeRelativeToShorterEdge(const cv::Mat &original_image, const float relative_size, cv::Mat &destination_image)
+{
+    const float target_edge_length{std::min(original_image.cols, original_image.rows) * relative_size};
+    cv::resize(destination_image, destination_image, cv::Size(target_edge_length, target_edge_length));
+}
+
+cv::Mat createFullSaturationFullValueAlphaMaskOf(const cv::Mat &img)
+{
+    cv::Mat mask{};
+    cv::Mat img_hsv{};
+    cv::cvtColor(img, img_hsv, cv::COLOR_RGB2HSV);
+    cv::inRange(img_hsv, cv::Scalar(0, 254, 254), cv::Scalar(180, 255, 255), mask);
+    cv::cvtColor(mask, mask, cv::COLOR_GRAY2RGB);
+    return mask;
+}
+
+void blendIntoBottomRightCorner(cv::Mat &img_to_blend_in, cv::Mat &img)
+{
+    const cv::Rect roi = cv::Rect(img.cols - img_to_blend_in.cols, img.rows - img_to_blend_in.rows,
+                                  img_to_blend_in.cols, img_to_blend_in.rows);
     cv::Mat img_roi = img(roi);
-    // cv::Mat my_mask{};
-    // cv::threshold(img_to_blend_in, my_mask, cv::Scalar())
-    cv::multiply(cv::Scalar::all(1.0) - alpha_mask_of_img_to_blend_in, img_roi, img_roi);
+    cv::Mat mask = createFullSaturationFullValueAlphaMaskOf(img_to_blend_in);
+    cv::multiply(cv::Scalar::all(1.0) - mask, img_roi, img_roi);
     cv::addWeighted(img_roi, 1.0F, img_to_blend_in, 1.0F, 0.0, img_roi);
 }
 
-void addColorWheelToLowerRight(cv::Mat &img, const float relative_size = 0.2, const size_t hue_offset = 0)
+void addColorWheelToBottomRightCorner(cv::Mat &img, const float relative_size = 0.2, const size_t hue_offset = 0)
 {
-    cv::Mat color_wheel_rgb{};
-    cv::Mat color_wheel_mask{};
-    createColorWheelImageAndMask(color_wheel_rgb, color_wheel_mask, hue_offset);
-
-    resizeImagesRelativeToShorterEdge(img, relative_size, {color_wheel_rgb, color_wheel_mask});
-
-    blendIntoImage(img, color_wheel_rgb, color_wheel_mask);
+    cv::Mat color_wheel = createColorWheel(hue_offset);
+    resizeRelativeToShorterEdge(img, relative_size, color_wheel);
+    blendIntoBottomRightCorner(color_wheel, img);
 }
 
 int main(int argc, const char **argv)
@@ -317,7 +333,7 @@ int main(int argc, const char **argv)
 
     const size_t edge_img = 800;
     cv::Mat img(edge_img, 1.5 * edge_img, CV_8UC3, cv::Scalar(100, 100, 100));
-    addColorWheelToLowerRight(img, 0.15, 0);
+    addColorWheelToBottomRightCorner(img, 0.15, 0);
     cv::imshow("Added images", img);
 
     cv::waitKey(0);