Show a color wheel indicating actor direction in the grid images, as advertised in the README: 

![dogm_iter-7](https://user-images.githubusercontent.com/27212661/78576637-39fd8980-782d-11ea-96a1-15ecf20bb26e.png)

Features:
- Wheel can be rotated with an integer parameter (orientation is currently correct)
- Wheel size is relative to original image, given by a float parameter

Considerations:
- I just put all this into main.cpp. I think we can devise a better structure for the new functions and for other free functions in `main.cpp`. I'm thinking of a utils file. If you agree, I could create a PR with a suggestion for that.