@TheCodez 2020-04-06
This is awesome!

Yes that would be nice. If you want you can also extract the simulator class into its own `simulator.h` file. 

Thank you for all the great work so far 👍