@cbachhuber 2020-04-07
Thank you very much for the kind words! So far, I'm having fun here, also thanks to your support.

Structure sounds good, I'll tackle that in the coming days.