@TheCodez 2020-04-07
Out of curiosity, what execution times do you get on which type of GPU?