@cbachhuber 2020-04-07
Good question! I'm on a Nvidia Quadro P2000, driver 435.21. I think that performance-wise, it's roughly equivalent to a GTX 1050. I see the following execution times of `demo` on the current master branch:

| Parameter modification  | Cycle time | Setup duration (until the for loop in main) | overall execution time |
|-------------------------|------------|---------------------------------------------|------------------------|
| none                    | 138-156ms  | 44s                                         | 47s                    |
| particle_count = 1*10e5 | 52-63ms    | 12s                                         | 13s                    |
| particle_count = 6*10e4 | 35-39ms    | 6.8s                                        | 8.4s                   |

I don't see a significant time difference between release and debug, tested with a few runs. I saw a strong influence of `particle_count` on the execution time of [setupRandomStatesKernel](https://github.com/TheCodez/dynamic-occupancy-grid-map/blob/master/dogm/src/kernel/init.cu#L35) when playing around with the code earlier. Is the same true for you? Which executions times do you see? Also, is such a high particle count necessary? When I used lower numbers (>5*10e4), I didn't see (subjectively) worse results.

Since execution time is one of the main motivations of having this project in addition to [mitkina/dogma](https://github.com/mitkina/dogma), I would suggest documenting mine and your cycle/iteration times somewhere quickly accessible (in the top-level readme, or in a file linked from it). What do you think?