@cbachhuber 2020-04-08
I'm looking forward to your setup time improvement!

For the paper, they use a GTX 980, which is [more than twice as powerful](https://gpu.userbenchmark.com/Compare/Nvidia-GTX-980-vs-Nvidia-GTX-1050/2576vs3650) as our GPUs. Therefore, if you achieve approximately double the cycle time as reported in the paper, you should have an equally optimized algorithm, right? Of course there is still some way to go.

I see, so the particle count is well motivated :+1: 

Cool; I would also mention the GPU performance difference between the paper and our experiments.