@aquac 2020-04-09
@TheCodez @cbachhuber Sorry for hijacking this PR, but I have a question regarding performance and memory demands I didn't want to create a new issue for that.

Given N_p persistent and N_b birth particles, what is the max. required memory for this implementation?