@cbachhuber 2020-04-09
I quickly checked this with `watch -n 0.2 'nvidia-smi'` while letting the demo run, and I see that the demo application requires 130-220MB on the GPU. 
This is on the current master, so `N_p=3E5` and `N_b=3E4`.

@TheCodez can you confirm?