@TheCodez 2020-04-10
I need around 290MB using the same numbers. 
One should also note that the memory is also highly dependant on the grid size not just the particles. In fact a particle only is 28 bytes whereas a gric cell needs 60 bytes.