@aquac 2020-04-10
@TheCodez @cbachhuber  thanks a lot!
One more question, since I didn't dig into the code, is the maximum memory demand determined by (N_p + N_b) or just N_p because the old particles are discarded beforehand?