@TheCodez 2020-04-10
@aquac by (N_p + N_b), they are used until the end and never discarded.